[utilities] Remove String#escapeHTML and String#unescapeHTML
String#unescapeHTML is never used so it is safe to remove it.

String#escapeHTML has only one use-site. The ProfilesPanel uses it to
escape the title of a tree element. This title is than later assigned
to a {textContent} property resulting in the pseudo code:

    element.textContent = title.escapeHTML();

which causes the HTML escaping to actually happen twice. This means
that String#escapeHTML is safe to remove as well.

Change-Id: I7f55d12c0e26fa7acc595e15f63d34dbb13bbc8d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2046547
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>