High contrast work for treeoutline.
Centralize common styles for hover, selected and focus. This change
makes sure the text and icons do not change color when hovering on an
element that is not selected. Text and icons will change color on
focus and selected. The color on the background property will
change when focused and selected.

Bug: 1048378
Change-Id: I574e5fc921045a836546c224ac31e38d8eb3f8ff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2049039
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Kham Udom <khamu@microsoft.com>