High contrast color fixed to toggle buttons, inputs and opacity
The patch fixes the toggle buttons for animation speed and the box
shadow type, inside the shadow editor.
Set opacity on disabled color. They are dimmed in places like the
text prompt in the Coverage panel and inputs and button in the
Sensors panel.
Fixed the hover state in the Request blocking panel. Where the
hover gradient is removed and background is set to Highlight.

Before: https://imgur.com/a/JHAzhS5
After: https://imgur.com/a/RTbiYad

Bug: 1048378
Change-Id: I5996548ee43dd8e866e466860dcf601d232f9e5c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2070664
Commit-Queue: Kham Udom <khamu@microsoft.com>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>