[e2e] Add user metrics tests
This migrates the user metrics layout test from Chromium. A future CL
will deprecate the layout test in Chromium in favor of this test. In
order to avoid tracking internals of DevTools this CL augments the user
metrics class to dispatch DOM events on Window with the payloads for the
purpose of testing. This decoupling means that if the internal
implementation changes the tests can continue to run (provided the
events are dispatched).

Bug: 1044632
Change-Id: Ie85230451aa373bb09f3288045a466bab862d6e8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2071885
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>