Remove tsbuildinfo files from ts_library outputs
The Chromium deterministic builder has been failing since https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2067062
has been submitted. It appears to be failing on the `.tsbuildinfo` contents, which we suspect is related to
inconcistent key ordering in the json file. However, these files
were generated for the DevTools source files from the start, but it only started failing after a `ts_library` was added
for a test.

We initially thought that the `.tsbuildinfo` file contains absolute paths, which would trip up the deterministic builder. However,
after investigation this was not the case. Additionally, the `.tsbuildinfo` files aren't an input for any Ninja rule in DevTools.

We have tried to determine why the builder is erroneously including these files, but without success. Therefore,
we will apply the interim solution of removing the `.tsbuildinfo` files from the outputs to unblock the deterministic builder.

Later, we should figure out why these files are included and also what the actual diff is from `compare_build_artifacts`. (The
latter only compares file size and does not provide an actual diff)

Bug: 1053887
Change-Id: Ia9f6753a557de40028c84a48381fbf006edf4318
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2074757
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>