Fix clicking on labels for checkboxes
This was a regression introduced in
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2062603
that broke the clicking of the labels corresponding to the checkboxes.

To fix that, we revert the changes to the labels, but keep the
`aria-label`. The label is necessary, as the label and input are in
shadow dom and otherwise do not show the correct descriptive name.

Fixed: 1056324
Change-Id: Ia34210961bcb7f5927dc3663633e27ed82b363ff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2077666
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>