Select main thread in memory tab by default
The main thread should be selected in memory tab. But sometimes other threads like the ServiceWorker one are selected instead.
Fix - By default first item in the list is marked as selected. But we have to mark the item with id `main` as selected.

Fixed: chromium:1044767
Change-Id: I612f7ce37edc233fd1c24a683008739f5a627d8b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2080380
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>
Commit-Queue: Jan Scheffler <janscheffler@chromium.org>