Add tests to verify sources panel shows lazy-loaded files
This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/dynamic-scripts.js;drc=77578ccb4082ae20a9326d9e673225f1189ebb63?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest

It updates the server to properly serve CSS files. It also updates the
sources-helpers to decouple the "opening the sources tab" and "opening a
source file in the code editor". This allows the new test to only use
the former and not the latter.

R=aerotwist@chromium.org

Bug: 1044632
Change-Id: I22597454117a5ae05a7bcd541389a9ce1424230e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2083316
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>