Add skipped test for breakpoints with inline scripts
This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-inline-with-sourceURL.js;l=6;drc=45107a2b20f840bd1833c33ffdd6383fb56de32d?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest

Note that this e2etest is currently skipped because of crbug.com/1003497
as this functionality is broken. The layout test is erroneously passing
and the e2etest is catching this issue.

R=aerotwist@chromium.org,sigurds@chromium.org

Bug: 1003497,1044632
Change-Id: Id51d1ad849cd978b94731dc5f6bb6fccb227a8b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2084583
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>