[workspaces] Don't show a console error when file select is cancelled
The backend will pass through a different value for when the selection
was cancelled, instead of a failed selection.

We used to pop out the console for this, which was confusing. Now we do
nothing in the UI when the selection is cancelled.

Chrome side change:
https://chromium-review.googlesource.com/c/chromium/src/+/2089837/

Bug: chromium:787728
Change-Id: I30f282530461ad168d5291d1eb88fdc02f2ac9cf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2089898
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>