Prepare Linkifier.js for asynchronous live locations
The Linkifier is a widely used component. We don't want to asyncify
all the call-trees and since {await}ing the live locations here is not
strictly necessary, we simply delegate some actions to a then handler.

As the varous {createLiveLocation} methods do not yet return promises,
we use Promise.resolve() as a migration path, so we can turn
{createLiveLocation} async independently of touching the code in this CL.

R=mathias@chromium.org

Bug: chromium:1032016
Change-Id: I4709a327d1ab4772676254fa500c8ae35693d512
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2072384
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>