[utilities] Port Array#remove to a ArrayUtilities.removeElement helper
This CL is part of a cleanup effort to get rid of Array#remove. The
code for the helper and the unit test is duplicated for now. Once
all use-sites of Array#remove are cleaned up, we can remove
Array#remove itself (plus the unit test).

R=jacktfranklin@chromium.org

Bug: chromium:1050549
Change-Id: Icef15af1a85d15e2d833a5ac2ba95e3591bb256a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2094124
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>