Sort generated property values on keys
When dumping the JSON object, we have to ensure that the keys are in the
same order. Otherwise, the generated file is non-deterministic.

We hit this issue when we made a subsequent update in
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2089908
where we are adding an additional dictionary to the file, causing
failures on Windows.

R=changhaohan@chromium.org

Bug: 1039620
Change-Id: Iec33962d4fd5dff461bc2930db54e5c85ce645b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2095300
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>