[utilities] Replace Array#remove with helper in WatchExpressionsSidebar
This CL is part of a cleanup effort to get rid of Array#remove. Please
note that we should actually convert {_watchExpressions} to a Set,
but as a web tests directly accesses this variable a 3-way change is
needed. To unblock the cleanup project we replace the single use-site
with a helper call and leave a TODO.

R=jacktfranklin@chromium.org

Bug: chromium:1050549
Change-Id: I16ddfae95bb88dbfa082ce3bc51ec737ede61548
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2095310
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>