[network] Clicking on 'View Headers' actually opens the 'Headers' tab
This CL extends the RequestActivated event by an optional Tab for the
NetworkItemView. For the 'View Headers' link we want to show the
'Headers' tab of the NetworkItemView.

Please note that we can't use the Revelaer for NetworkRequests as
the revealer requires an already open NetworkItemView which is not
the case for this workflow (looking at requests in the log view).

R=sigurds@chromium.org

Bug: chromium:1060136
Change-Id: I868e03dfbc6086d44a60db3d63e2199eb17a3b82
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2097982
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>