[ui] Add spacing to checkboxes in lighthouse start page
The other forms with the checkboxes leave the default margin on the
checkbox elements but we remove it from the radio buttons for some
reason. This CL therefore also moves the radio button in 3px by restoring
the original margin-left, which matches the indent (relative to the
section title) of the other forms on the page.

Remove the '.lighthouse-form-section input rule' because this was not
applied to any of the other forms like checkboxes as they are in
shadow roots.

Move the radio button text into a span and add margin to it explicitly
which matches what we do for the checkbox text.

Before: https://imgur.com/a/cJaPwqU
After: https://imgur.com/a/X3ol0kL

Bug: chromium:1061362
Change-Id: Ia718758304450296a9a5c9fbdb8bf15b477b4324
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2100916
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>