Add coverage reporting to ninja typescript
This post-processes all JavaScript files with the Istanbul instrumenter.
The files in question are the output files from TypeScript, which means
that Istanbul operates on the TypeScript sourcemaps. As such, the coverage
report will have the correct lines.

For an unknown reason, Istanbul still adds empty files to the coverage report,
in the `out/Default/gen` directory. Since normal files are always loaded,
we ignore files with no coverage. It would be great to later figure out why
that is happening, but for now we are happy with skipping them.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1061125
Change-Id: I754e8b9f46dd2fdb167eda6e64927c94b88ff676
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2107218
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>