[issues] Show reload InfoBar when issues might be incomplete
For complete and accurate issues, DevTools needs to be active while a
page is loaded. Should the {IssuesModel} not have observed such a
corresponding navigation event, we notify the user that a reload might
be a good course of action.

Screenshot: https://imgur.com/a/3aCDwV2

R=sigurds@chromium.org

Bug: chromium:1060628
Change-Id: I61d5deb1adfe71b392cd25571ce771d820c7b012
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2116440
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>