Disable popover eager evaluation on Wasm frames.
For now, we neither have the front-end nor the back-end logic to perform
Debug-Evaluate of Wasm code, and the popover evaluation is just
confusing since it tries to treat the Wasm as JavaScript. So disable
that logic for now.

Bug: chromium:1063875
Change-Id: I2ab923076b076f55a5318b2f15b10cd921e81e6d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2119688
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>