[issues] Clicking on affected requests will open the details view
Instead of just highlighting the affected request in the network panel,
clicking the link now opens the detailed view for a request in the
network panel.

This CL adds a new utility method on the {NetworkPanel} that ties
together existing functionality. Currently, clicking an affected
request opens the "Headers" tab of the request. This is done because
the "Affected Requests" views is re-used for all issues where requests
are attached, not just cookie issues. In the future, we might be able
to pass along some data, so we could choose the opened tab depending
whether its a cookie or a COEP issue.

R=sigurds@chromium.org

Change-Id: I1118b2e2c428c6135f895816d1f7fdaf4d4ce79b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2123481
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>