Focus default element after showing Settings screen
This CL fixes a focus regression with the Settings discoverability
changes [1] where the default element of a Settings view (tab) is not
focused when externally revealed.

The regression was caused by the replacement of UI.viewManager.showView
for this._tabbedLocation.tabbedPane().selectTab which does not call
focus() after completing.

Ths fix is to call .focus() on the tabbedPane element, which will
then call .focus() on the underlying View widget, which focuses the
default element of the View.

Demo: https://i.imgur.com/yUKxPa5.gif

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2049183

Bug: 1069337
Change-Id: If2f038388fe21b781560e380c28e483a1438d79c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2144433
Reviewed-by: Vidal Diazleal <vidorteg@microsoft.com>
Reviewed-by: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Commit-Queue: Brian Cui <brcui@microsoft.com>