Change waitForFunction condition to add more debugging output in network-datagrid.ts
Even with 100 iterations, I am unable to make this test flake as was reported in
crbug.com/1066813. Since the error message does not show any information regarding
what the state of the application was that was incorrect, it is not possible to
understand what went wrong.

Therefore, update the waitForFunction condition to wait for at least 3 cells.
The current assumption is that the test in question had more than 3 cells.
By updating the condition, we can have a more descriptive assertion failure
in the `assert.deepEqual` below it, hopefully pointing us to the actual culprit.

R=aerotwist@chromium.org

Bug: 1066813
Change-Id: I89916a5a963cf9ae6be350e7ca29e438dafb174f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2139973
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>