Return an error in evaluateVariables when formatters are unavailable
Generating the evaluator module when the formatters lib is not built is
not actually an error, but will generate a module with a bunch of
additional imports. That's intentional, but leads to strange errors in
the protocol clients. This CL adds a better error.

Bug: chromium:1020120
Change-Id: Ie3a77e4c5a0b8371fffb7233b130092800f4a8df
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2183920
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>