Devtools: remove tabIndex when JS VM instance list is empty
Issue: On Profiler, the Javascript VM instance list has a tabIndex 0 even the list is empty, which results focus lost alike behavior when traversing using Tab key.

This change sets tabIndex to 0 if the list is not empty and resets it to -1 when it becomes empty.

Bug: 1078243
Change-Id: I265656f0cdfe7c412ceb37cef93f5fb33873df48
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2199400
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>
Commit-Queue: Junyi Xiao <juxiao@microsoft.com>