Fix: Memory panel e2e test.
Fix aria label and wait for sidebar content.

We were previously waiting for the enclosing panel which did not guaranteed that the button existed. This caused flakyness..
This CL waits for the sidebar panel content to be loaded before looking for the button to take a new snapshot.

Bug: 1083134
Change-Id: I6f66df6b2f82deb0075360427985b6447d730166
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2204361
Reviewed-by: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jose Leal <joselea@microsoft.com>