Add Dispatcher base-interface to enforce type-safety
To make sure that typescriptified dispatchers use the object
notation, we have to signal to the InspectorBackend that
we use the object notation. As such, add a base interface
with an interface method called `usesObjectNotation` to make
sure that all dispatcher implementations use this method.

A follow-up CL will make the required changes to the
InspectorBackend to call the dispatcher with the correct
method arguments.

R=jacktfranklin@chromium.org,aerotwist@chromium.org
CC=sigurds@chromium.org,janscheffler@chromium.org

Bug: 1081686
Change-Id: I1fa9338afa44d26cc63dce84b9d88a936c4f35e4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2207236
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>