Use terser rollup plugin
Terser can minify variables names and is a smarter minifier than
the previous regex-based minifier. The regex-based minified only
removed whitespace and comments.

This reduces the total size of resources loaded for a fresh load
of the Elements panel from 3.4MB to 2.6MB.

It does increase the build time for a release build, but this should
not have an impact on the development workflow of engineers working
on DevTools.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Fixed: 1084578
Change-Id: I037206da9fe2f3e6148fd814f5af7db78d543cee
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2208139
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>