Typescriptify HeapProfilerModel
This discovered a mismatch of return types; the HeapProfilerModel leaves
out non-optional fields for source location information, and apparently,
the code that displays this information can deal with the absence.
This CL introduces a new super type for the different profiles, which
also allows undefined for some fields.

Bug: chromium:1011811, chromium:1079233
Change-Id: Ic0a4625f435b73f85fc29765d37fe74aeece4b69
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2204111
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>