[sdk] Properly encapsulate and await `Script` content.
The content of the `Script` is lazily loaded, but wasn't properly
encapsulated in a Promise and `await`ed correctly everywhere. This
change fixes that and removes the `_source` and `_originalSource`
fields from the `Script`. There are still data races left with
respect to Wasm source mapping, and those will be fixed in a later
CL once it's clear how we propagate this information.

Bug: chromium:1056632
Change-Id: I43b47825cfa41a8595c353b4205c0ef471ed524d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2218008
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>