Add DOM helpers for testing components
This change is pulled out from the ElementsBreadcrumbs CL and introduces
some basic DOM helpers to help with rendering elements for tests and
cleaning up the DOM afterwards. I expect as we write components and test
them these will grow and change but this is a good starting point.

Change-Id: Icd3f3d4132ac75a62d6df43382e2d3e46cdcb580
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2230501
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>