Remove style three-dot menu in favor of autocompletions
The style panes have a three-dot hover menu that would show buttons
to add a strict set of properties, mostly related to colors. The
primary use case for these buttons would be to provide the correct
syntax for the properties.

However, we can use the auto completions for this. With this change,
typing for example `box-shadow`, the auto completions provide the
"0 0 black" value. This also improves the UX for users who use the
auto completions feature solely (which is the majority of developers).

Before: https://imgur.com/a/VLQ8goH
After: https://imgur.com/a/OLNWeYa

R=petermueller@chromium.org,raweil@microsoft.com

Change-Id: Iee262d31ced51ed019cfa9af05d0155ff943b306
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2232549
Reviewed-by: Peter Müller <petermueller@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>