Fix score function for filenames with dots or whitespaces
Previously words in filenames that are seperated by dots or whitespaces could receive a lower score as string match in the path string. This can lead to a folder in the path being highlighted when searching, while not highlighting the matching filename.
For example in "App/my.App.js" the App-folder would be highlighted, but not the filename, when querying for "App".
This CL approaches this by adding checks for both dots and whitespaces to the testWordStart function that is used to calculate the score.
To evaluate the results easier, unittests have been added.

Bug: 1084494
Change-Id: Ia7a0ea3ed7ed033bb6bd35eb00f2ae75a0046afe
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2218016
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>