DevTools: E2E test covering going from perf to sources
Simple test in the existing workflow_tests.ts file to make sure
this user flow works.
A similar CL was landed and then reverted before, as this test is just
a tad slow and will often timeout.
This new CL removes a single waitFor in the performance-helpers.ts file
in an attempt to make the test run a little faster.
It also increases the timeout for this test suite to 5s instead of 3s.
Indeed, moving from panel to panel takes time in DevTools and 3s was
just not enough for the test to pass every time.

Bug: 1079076
Change-Id: Ib6e31e943ccbddc9d82aea5da394c371499c1aab
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2230774
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>