Show min- and max-width media inspector labels in the correct order
Currently, min- and max-width labels in media inspector markers are
shown incorrectly on the left side, with the max-width on the inside.

This change swaps the labels so that they are shown in the correct
order on the left side.

Before: https://i.imgur.com/8egqiV1.png
After: https://i.imgur.com/CYAZd80.png

Bug: 952695
Change-Id: I06f7db2fa345a007dce2ebc3b3d6f44b7845f43c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2238227
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>