DevTools: Improve setPanelIcon logic
The code made the assumption that the panels were always found in the same tabbed location.
This is no longer true, since we introduced movable tabs.

With this change the code now looks for the correct tabbed location and handles the case where the panel is not found.

Bug: 1093507
Change-Id: I7ae1e09ce0059ae5a99ded3b6764dc5cb22455a0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2239113
Commit-Queue: Jose Leal <joselea@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>