Grid: negative grid line number tooltips
This change adds tooltips to the Grid overlay to show negative line
numbers, when those are present in the data sent by the backend,
and when configured to do so.
The default configuration does not show them yet.

With positive and negative numbers: https://i.imgur.com/riUXwz4.jpg
There are still some details to address when labels overlap eachother
at the corners, but we should do this in a follow-up.

Labels do flip the other way when too close to the edge:
https://i.imgur.com/W9iACIr.jpg

This CL has a corresponding chromium.src CL: https://chromium-review.googlesource.com/c/chromium/src/+/2242952/

Bug: 1047356
Change-Id: I072579c2db7afbd7683ffbe5cf00cca4ad16a73a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2239367
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>