Update Settings telemetry panel keys to fix Locations tab
This tiny CL fixes a Settings panel code naming mismatch that was
introduced by [1] when the Settings Locations tab id was renamed from
"emulation-geolocations" to "emulation-locations".

This caused panel-open actions to stop flowing for the Locations tab.
This CL simply updates the panel code mapping to refer to the
new panel id, so the panel code is valid again.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2119672

Bug: 1096745
Change-Id: I0de65d6dd3160bbf20deb8fce9049075863f68fe
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2252472
Reviewed-by: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Jose Leal <joselea@microsoft.com>
Commit-Queue: Brian Cui <brcui@microsoft.com>