[devtool] Change "FPS meter" to "Frame Rendering Stats"
With this CL landed:
https://chromium-review.googlesource.com/c/chromium/src/+/2244002

The fps meter now have a completely new look, and it is actually
showing the throughput of the page. Other than that, it also
shows other diagnostic information about frame rendering besides
just the FPS.

This CL changes the name of "FPS meter" to "Frame Rendering Stats".
Its description is also updated.

The corresponding CL for chromium is:
https://chromium-review.googlesource.com/c/chromium/src/+/2268398

Bug: None
Change-Id: I02a7348eca9fd19c5433d9be455c0c811070c91b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2269939
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>