Grid: Draw grid area names
This is based on this backend change:
https://chromium-review.googlesource.com/c/chromium/src/+/2275484

When the backend returns data about the named grid areas (their names
and paths), then the frontend of the overlay needs to draw them.
The mockup for how they should look can be found here:
https://docs.google.com/presentation/d/1krTiCT_c5sIAXBU6hpSok9ev3cZDRyucHf6rnKNTa8o/edit#slide=id.g74d98c0ef3_1_26

Grid areas are basically just highlighted with a thick border around
them, and the name is displayed in the top-left corner.
This change does not handle cases where multiple areas overlap,
resulting in overlapping names. This will be handled later.

This change refactors a little bit the code in css_grid_label_helpers.js
to make it easy to generate more label types in the future.
As before, whenever a grid is drawn, all labels are cleared. This hasn't
changed.
Line number labels and area labels are now in 2 different DOM containers.

A new set of unit test cases were added too.

And finally, a new option in the settings panel was added to turn areas
on or off.

Screenshot: https://i.imgur.com/NVqAUsn.jpg

Bug: 1100384
Change-Id: Icb935a1a6e143bfdd95babb3c66a8f1412c80418
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2277595
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>