Use devtools_entrypoint for elements/
This allows us to introduce the bridges files as imports into elements/
which uses Rollup and terser. This makes sure the minifier will not
mangle away the HTML template tags, as well as properly minify the
content of the TypeScript implementations of a custom element.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1011811, 1061037, 1101738
Change-Id: Iaa1318c20f6d37c5aad97955f48ac16dacef246d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2277994
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>