Remove unused getter on Cookie
The port attribute was only allowed for Set-Cookie2, but we never
surfaced that information in our structured cookie display logic.

From a UX perspective, the port is part of the domain, and should be
communicated as such, so it doesn't make a lot of sense to keep this
around.

Fixed: chromium:1060613
Cq-Exclude-Trybots: luci.devtools-frontend.try:devtools_frontend_mac_rel
Change-Id: Ia4c23d47604185f55ef99377d5af2e7d073d97cd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2279842
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>