Extract copy to resources into separate template
We will use this template later in a different `devtools_module` template.
It also clearly documents how the remapping works to `resources/inspector`.

Ideally, we will remove this copy-step altogether, but since not all
files for DevTools live in `gen/`, we can't make that change yet. But now
we have a central place where we do the copying, so hopefully it is easier
to remove that once we are there.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1098730
Change-Id: I3981c97f99b3e3309ebe6ea0f4ec05fb0340835a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2282808
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>