Stop building formatter_worker-entrypoint with build_release_applications
This moves the formatter_worker-entrypoint out of the
build_release_application inputs. That's because this module is now
standalone and fully typechecked by TypeScript.

This also removes its legacy `module.json` support and stops using
the RuntimeInstantiator to start the entrypoint. That's possible,
because `formatter_worker` only relies on JavaScript files, which
the static imports already perform the loading for. The only
replacement is the `self.postMessage` which was previously performed
by the `startWorker` function in `RuntimeInstantiator`.

R=aerotwist@chromium.org

Change-Id: Icd18c0f16b42b9246ea138a9a68cbb06fd7e192b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2302603
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>