Add ad status to frame details in application panel
This builds on the back-end CL https://crrev.com/c/2289850,
which adds some basic reporting of the ad status of a frame.
We expect to improve the reporting such that the cause for
a frame being tagged as an ad can be debugged, but we will
do this incrementally.

Screenshot: https://imgur.com/a/pm4khn8
Bug: chromium:1095617
Change-Id: I74597c94ca9b6ece4646e0e551d45950462bec2a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2306152
Reviewed-by: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>