[wasm debug] Implement auto-stepping with the experiment.
This ports the auto-stepping logic that was already available with the
DWARF prototype to the `wasmDWARFDebugging` experiment. The condition
here is as follows:

  If we break on a raw location for which we have a reverse mapping
  to a UI location, then we check whether any of the raw locations
  to which the UI location maps is exactly the same as the break
  location.

This seems to yield the expected behavior, but comes at a quite
significant cost (transfering all the stack frames and scope chains to
the DevTools front-end just to decide that we don't want to break here),
so eventually we want to find a better and more efficient way to do
this.

Bug: 1105765, 1083146
Also-By: pfaffe@chromium.org
Change-Id: Idc4a4f57f352322769557609dde7fe2aa26a7ce2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2300115
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>