Grid: Display grid line names in the overlay
This is based on the following backend CL:
https://chromium-review.googlesource.com/c/chromium/src/+/2291435

This change makes the frontend of the overlay display line names when they are
present in the data sent by the backend.
This is a first step only. More work will be needed to avoid overlapping labels
in some situations.
If both numbers and names are returned, then only names are shown, to avoid
overlaps. The UI spec doesn't require us to show both at the same time for now
anyway.
To avoid complex edge cases for now, the number of names per line is limited to
3, and there's a max-width applied to names too (with ellipsis).

The goal is to try this out for real, and then iterate on the design based on
what we learn.

Screenshot: https://i.imgur.com/RWQCg0h.jpg

Bug: 1100385
Change-Id: Iffd4d3f1f732a5583e7cf007ae9f83a7ffa3dda2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2292205
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>