Aria-label the scope entries
This CL assigns an aria-label to list entries in the scope view. The
label is attached to the actual list item and not the span inside of
that whose textContent contains the name currently. The textContent
sometimes is just the name and sometimes both the name and the type.

Change-Id: Ia4f916057ed9c049c7b72537822ac123149f13d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2325216
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>