Remove reload banner in the issues tab
After https://crrev.com/c/2332229, all issues are generated based on
isses reported from the back-end. This means that the prompt to reload
is not as urgent anymore (as all issues will be showing, but some may
not be able to link to the network request, as the network log might be
empty).

Bug: chromium:1072335
Change-Id: I34cc45fbe3a36bb5598b7a08e3848a9c549bbc13
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2334967
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>