Grid: Don't assume grid lines are straight when drawing extension lines
A grid container can be rotated or skewed using CSS transforms. It can
also be drawn in a different writing direction.
Because of this we can't assume that row and column lines are straight
or that rows are horizontal and columns vertical.

This change changes the approach with which we've been drawing extended
lines in the overlay. Now, we look at the 2 end points for each line,
and based on this alone we draw the extended lines. This is done by
computing a and b in y=ax+b based on the 2 points we have.

Before: https://i.imgur.com/exTaDBo.jpg
After: https://i.imgur.com/fnzrOFw.jpg

Bug: 1112363
Change-Id: I2f874c3f8d4de81ce0b5048c1bd585264a93e10a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2334363
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>