[issues] Re-introduce issues tab toolbar and third-party checkbox
This CL re-enables the toolbar for the issues tab and adds a checkbox
to allow users to enable/disable inclusion of third-party issues
in the list. The default will off, so that issues caused by
third parties are not shown.

Screenshots: https://imgur.com/a/weQBIRd

R=petermueller@chromium.org, sigurds@chromium.org

Bug: chromium:1080589
Change-Id: I9bb85c420f356e9e6da89d71e01ed964a645dc12
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2339136
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Müller <petermueller@chromium.org>