Grid: Fix line name label positioning regression
Recently, in CL https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2327714
we made changes to the format of data sent by the backend to the overlay
frontend.
Before this change, we used to send single-value offsets to position
the labels inside the grid. These values would then be added top the
known grid bounds to determine their actual absolute values.

We moved away from it because when a grid is transformed, it may be
that the labels to be positioned aren't aligned horizontally or
vertically with the bounds of the grid.
We now use x/y coordinates to position the labels in the overlay
frontend.

However in doing so, we forgot to update the code in
_normalizePositionData that deals with line name labels. The
position for these names are now being sent in x/y form, but the
frontend code was still expecting them as single-value offsets. This led
to NaN values when calculating their positions.

This change fixes this, and adds a unit test to make sure this does not
occur again in the future.

Before: https://imgur.com/D8Ac7Vz.jpg
After: https://imgur.com/7i88N20.jpg

Bug: 1100385
Change-Id: I8529a05d15d99ecdd88feb9efd2adf8702e8d946
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2340315
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>