Elements: Remove Event Listener Pane focus override
This CL fixes an apparent focus issue in the Elements Event Listener
sidebar pane. When the event listener tree outline was made
keyboard accessible, it was given focus by default when navigating
to the event listener pane from a mouse click.
However, since the tree was not keyboard
focused or marked focus-visible, it appears that the Event Listener
tab is still focus.

The fix removes the focus method from the widget, which causes focus
to remain on the tabbed pane header, which was the previous behavior.

This does not change the keyboard navigation, as keyboard only users
would still use arrows to nav the Event Listener
tab pane and then tab into it. This is the same behavior as the
Styles pane.

Before video in the bug
After: https://imgur.com/2g7W9MC

Bug: 1114604
Change-Id: I103dccef37bfe605e6093bb0f0d4104e3670da14
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2351165
Reviewed-by: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Jose Leal <joselea@microsoft.com>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>