Force a screen reader pause between a tag name and attributes
When focused on a <div id="..."> element in the Elements panel, a screen
reader will announce "divid", which is confusing.

This change adds an aria-label to the tag name  part of the previewed
element. This forces screen readers to announce the tag name separately
from the attributes.
With this change, screen readers will announce "div", then pause, then
announce "id".

Worth noting though: the word "div" will be spelled out (at least on
Windows Narrator) as "D" "I" "V". I guess that's expected as div is not
a word. Other tags such as span, body, script will be read as normal
words. However I wouldn't expect that to be the case for the tag html,
so I'm inclined to think that spelling out div too isn't a problem.

Bug: 1115449
Change-Id: I479f38c2db6ea5c068ecacf4c9a482f8cc8168a3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2351163
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Jack Lynch <jalyn@microsoft.com>