Fix blur handler of BreakPointEditDialog
When the user clicks on the `this.element` div, then the `blur`
handler of `this.contentElement` fires as it loses focus. Inside
the blur handler, we look at the element that will receive focus,
i.e. `relatedTarget`. The handler compares the `relatedTarget`
to `this.element`, not dismissing the dialog.

Since the `blur` handler is installed on `this.contentElement`, it
will no longer fire, even if `this.element` loses focus, causing
the dialog to stay open.

The solution is to install the `blur` handler on `this.element`,
to stay consistent with the check inside the `blur` handler.

R=mathias@chromium.org

Bug: chromium:1081452
Change-Id: I84b01901110d0ce2acda088d3d8fc420afea6b7a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2381414
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>