Add filter for full request url in network tab
This CL adds a new url: filter to the network tab that filters
on the full URL, including the protocol information (e.g. https://)
in the search query.

Note on default filtering: Originally this CL modified the default
filtering by using the full request url, instead of the current
implementation which relies on the request path and name. Since this
change might have confused power-users and a more thorough rework
of the filtering is planned, the decision was made to add an additional
url: filter instead.

Bug: 1104188
Change-Id: Ibddd26eca7fb56eb9d08277df72c0da3acbb53a1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2318146
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>