Add css variable for link-color
This CL adds a variable for the link color in both light and dark mode.
This is in preparation for fixing the color of icons which are links in
the issues tab. But by removing the the yellowish devtools-link color
in the dark theme it already removes some inconsistencies.

before: https://i.imgur.com/ZcGhG9C.png
after: https://i.imgur.com/WxWpjYS.png

Bug: chromium:1122511

Change-Id: I17703643d8d0180f28761a6adf56f2b92ee1e4aa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2391225
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>