Refactor away externally adding property onto imported object
This is to prepare for ComputedStyleWidget.js's TypeScript check. This
CL removes the behavior that adds an internal property `_propertySymbol`
to an imported definition: UI.TreeOutline.TreeElement. This field
cannot and shouldn't be registered in TreeElement's definition, and
removing it will make TS-check much easier.

Change-Id: I4ac5018e145e6ca4c0c9aa5f59585840d290f31f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2391252
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>