[loc] Escape paths correctly
The localization_utils check a list of excluded directories against the
subdirectories of front_end. However, the list of directories hardcoded
the / separator, which is incorrect on Windows. This led to localization
checks in third_party subdirectories, causing build breakages and
performance regressions. This CL updates the paths to use the node
path.sep value, which in turn means that the string matching works on all\
platforms.

R=tvanderlippe@chromium.org

Change-Id: I5e1f2400b3f331ff4ac7acf953042b2ad900d40b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2398827
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>