Layout: Scroll to the grid container when the go-to-grid icon is clicked
In the Layout panel, each grid container in the list has an icon that
can be used to select the corresponding element in the Elements pane.

This change adds a secondary feature to this. On top of selecting the
element, the page will also be scrolled to reveal the element there as
well.

https://imgur.com/gPyIMLy.gif

Bug: 1109177
Change-Id: I6acfe12764f910341cc83a714d3a28a9e786bb55
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2404385
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>