Load accessibility_test_runner as module
The test runner was previously loaded as a script in the Runtime. To
be able to remove the scripts infrastructure from `module.json`, we have
to port all existing test runners to use `modules`. However, this means
that we are no longer including the test runner implementations in the
`_module.js` bundles that `build_release_applications` builds.

Therefore, we have to copy the sources ourselves to make them once again
available for loading. We update the `module.json` to use `modules`
instead, which then dynamically imports the file. We shouldn't use
`devtools_module` here, as we don't intend to include these files in the
GRD and should not end up in the eventual front_end bundle.

R=aerotwist@chromium.org,jacktfranklin@chromium.org

Bug: 1105476
Change-Id: I93380510198551d5b303d7ba136d04d0d1f2d4d9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2401163
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>