Grid: Use better colors for the grid overlays
Generating colors automatically was our first approach. Now we have a
list of colors provided by a designer. We want to use this.
In this change, we keep the same color generator class, but we just make
it cycle through these colors.
This change also slightly alters the label text color (while keeping the
logic to switch to white if it has better contrast, since we're going to
make colors customizable by users).

Bug: 1123052
Change-Id: I4b6606daa654d92e0c5d206e58e184b5578535b6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2408180
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>