Make it.repeat also do it.only
If you want to run a test multiple times, it's a safe assumption that
you don't care about the rest of the test suite while you're doing that.
So by default when someone uses it.repeat we should also isolate it with
it.only.

Change-Id: I71abba82c1c87a595924b4c12a1ed7af312b4bcf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2406161
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>