Restore Toolbar button focus background
This CL restores the button focus styles that were inadvertently removed
in another CL to fix an issue with the styles for buttons rendered as
links. The button styles here are identical to the old ones other than
a tweak to the selector to prevent it from conflicting with the
button-as-link styles.

Screencast with both types of button: https://i.imgur.com/j4pPI32.mp4

Bug: 1133075
Change-Id: I52de7607be04beb50f1331c3d441c9ae7cdc0ed5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2437771
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Jack Lynch <jalyn@microsoft.com>