Fix assertions of 0 mutations in component test
Sometimes when using `withMutations` it's useful to assert that some elements
have mutations but others do not:

```
await withMutations([
  // We expect one <tr> to be added
  { tagName: 'tr', max: 1},
  // but no buttons!
  { tagName: 'button', max: 0},
], ...
```

However, due to checking the truthyness of mutation.max, the `0` was always
overridden with the default value of 10. Swapping the check for an explicit
undefined avoids this.

Bug: 1129881
Change-Id: Id2317a35b590825cf2a8c07fc9d2c53cda200ea7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2436818
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>