Fix tabs shown on reload
The main toolbar containing tabs for the various panels contains a
counter for errors/warnings/issues/violations. This CL fixes the logic
for when this counter should be shown.

Previously, when there were violations but the violations experiment was
turned off, the counter would render a small visible element anyway.
This could lead to a bug where a page refresh would change the contents
of the main toolbar.

Fixed: chromium:1132775
Change-Id: I5bed294733267d204d12729efa3f3f7eb202d238
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2440894
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>