Network: Make the initiator column jump to the right position
The Network.Initiator CDP object now contains columnNumbers on top of
lineNumbers.
(CL https://chromium-review.googlesource.com/c/chromium/src/+/2442367)

When an HTML file is minified (all on 1 line) and loads a CSS file,
we need the Network panel to link from the initiator cell to the right
position in the Sources panel.

Before the CDP change, we would only get the line number, meaning that
the initiator would lead to line 1 of that HTML file. This is true, but
incomplete.
The CSS file might be loaded by a <link> tag at column 10000 on that
line.

This frontend change simply takes the new columnNumber info account and
that's enough to make the script formatter know where the initiator
really is in the pretty-printed file.
I also fixed a bug where we were using initiator.lineNumber as a
condition, which was wrong when that lineNumber was 0, since the
condition would be falsy.

Bug: 1134064
Change-Id: Id8426e8856194e14afcf8e4ec80943e6b461b4f3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2444236
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>