Use Node module resolution for tests
Our node tests run in the Node environment. As such, they already
include more type information, like Node, Mocha and Chai types.
However, it turns out that we did not update the moduleResolution
flag to enforce Node. Therefore, it was using the classic TypeScript
module resolution algorithm.

This wasn't a problem per se, as the classic algorithm will resolve
files in the same way as Node does. However, there is a slight
side-effect in that one of our `@types` we load references a type
that is specified in `@types/node`. The classic algorithm
erroneously traverses up the file tree to look for the filename,
in this case `stream`. If an engineer working on Chromium would
have a file called `stream.ts` in a parent directory of
`chromium/src`, then the classic algorithm would incorrectly resolve
to that file.

Instead, we should force the Node resolution algorithm, which
will not resolve to the `stream.ts` file, but will instead use
the types as defined in `@types/node`.

R=aerotwist@chromium.org

Fixed: 1135451
Bug: 1011811
Change-Id: I8ada319c923114f049f14fe38dcd8df82a9f2da8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2461774
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>