Settings: Prepare shortcut editor for reset button
This CL moves the population of the ShortcutListItem editor/display
elements into a new _update() function so that it can redraw itself as
necessary, rather than hanging on to references to certain elements so
that it can delete or re-append them. This simplifies the
addition/deletion of shortcuts and will make it possible to add a reset
button that reverts all shortcuts to their defaults.

There shouldn't be any user-facing changes from this.

Bug: 174309
Change-Id: I71666e1de9e24c93ed40518bfe05ea6f9a4ce4f4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2489815
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Jack Lynch <jalyn@microsoft.com>