Fix color-swatch regression in Sources panel
The color-swatch component was refactored in bug 1137438 and this
introduced a bug in the Sources panel where editing colors in
stylesheets using the color picker doesn't work anymore.

with this change, we now make sure the text for the swatch is available
as textContent, but hidden from the codemirror text editor, and we
re-wire the swatch-click event to show the picker.

Bug: 1141875
Change-Id: Idd811397ffaca9e5f4a2c54a7d6e2c163398c8c8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2495210
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Tony Ross <tross@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>