Adjust tabindex of network summary links to 0
This change allows a user to tab to network request links shown in the
summary panel of the Performance tab in devtools. This is done by
ensuring the tabindex of the element is set to 0, as opposed to -1.

Before: https://imgur.com/Bs12gVJ
After: https://imgur.com/qbbHWLq

Fixed: 1137785
Change-Id: Ia556376839e414df604a6f9727461d95cdd32017
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2486216
Commit-Queue: Jan Scheffler <janscheffler@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Jack Lynch <jalyn@microsoft.com>