Add setting for record network log action
Previously recording of the network log was enabled when reloading and not persisted by a setting.
This CL introduces a new setting which persists the existing action, so that it can be toggled in the network settings. Also the setting is not re-enabled when the page will reload.

Bug: 1122580
Change-Id: I537a1072f6d1dca6ef88f1b507a23977763341a4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2503513
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>